Combined Book and WelcomeMessage methods into MainMenu and removed redundant BookTest
