--- src/com/twu/biblioteca/Book.java
@@ -1,7 +1,0 @@
-package com.twu.biblioteca;
-
-public class Book {
-    public void showList() {
-        System.out.println("1. Clean Code | Robert C. Martin | 2008\n" + "2. Don't Make Me Think | Steve Krug | 2000\n" + "3. Test Driven Development | Kent Beck | 2000\n");
-    }
-}
