--- src/com/twu/biblioteca/WelcomeMessage.java
@@ -1,8 +1,0 @@
-package com.twu.biblioteca;
-
-public class WelcomeMessage {
-
-    public static void welcome() {
-        System.out.println("Welcome to Biblioteca. Your one-stop-shop for great book titles in Bangalore!");
-    }
-}
