--- test/com/twu/biblioteca/WelcomeMessageTest.java
@@ -1,23 +1,0 @@
-package com.twu.biblioteca;
-
-
-import org.junit.Test;
-
-import java.io.ByteArrayOutputStream;
-import java.io.PrintStream;
-
-import static org.junit.Assert.assertEquals;
-
-public class WelcomeMessageTest {
-
-    @Test
-    public void shouldWriteWelcomeMessage() {
-        ByteArrayOutputStream output = new ByteArrayOutputStream();
-        System.setOut(new PrintStream(output));
-
-        WelcomeMessage.welcome();
-
-        assertEquals("Welcome to Biblioteca. Your one-stop-shop for great book titles in Bangalore!" + System.getProperty("line.separator"), output.toString());
-    }
-}
-
